auth_service/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
sqlalchemy==1.4.27
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.8.2
pydantic[email]
python-dotenv
